Implement note persistence and a "Recent" notes screen
Introduces a local SQLite database to persist debit notes, allowing users to view, edit, and share previously generated notes.

- **Persistence Layer**: Added `NoteDatabase` (SQLite) and `RecentNotesRepository` to handle storage of notes as JSON strings (via new `NoteJson` encoder/decoder).
- **Recent Screen**: New `RecentScreen` with a tile-based list of saved notes, featuring a 90-day/500-row retention policy and empty state UI.
- **Note Management**: Added `NoteActionsSheet` for context actions on saved notes (Edit, Save a Copy, Share, Delete).
- **Navigation & Workflow**:
    - Saving a note now persists it to the database, resets the editor, and navigates to the Recent screen.
    - Editing a note from the Recent list re-stamps the date to today and treats it as a new draft.
- **UI/UX Enhancements**:
    - Replaced the Print action with Save and Share buttons.
    - Integrated `androidx.compose.material.icons.extended` for outlined icons (ReceiptLong, Schedule, Edit, etc.).
    - Improved Snackbar feedback with longer durations and success/error tones.

App renamed to "Debitter" in `strings.xml`.

+++ app/src/main/java/com/example/debitter/ui/components/AppSnackbar.kt
@@ -68,6 +68,6 @@ class AppSnackbarState(val hostState: SnackbarHostState) {
     private suspend fun show(message: String, tone: SnackTone) {
         this.tone = tone
         hostState.currentSnackbarData?.dismiss()
-        hostState.showSnackbar(duration = SnackbarDuration.Short, message = message)
+        hostState.showSnackbar(duration = SnackbarDuration.Long, message = message)
     }
 }
